int -> size_t in loops

+++ hd_solver.hpp
@@ -84,9 +84,9 @@ void lu_decomp(mdspan<double, dextents<size_t, 2>> a,
     std::vector<double> vv(a.extent(0));
 
     // fill in scaling vector
-    for (int i = 0; i <= ubound; ++i) {
+    for (size_t i = 0; i <= ubound; ++i) {
         double aamax = 0.;
-        for (int j = 0; j <= ubound; ++j) {
+        for (size_t j = 0; j <= ubound; ++j) {
             if (abs(a[i, j]) > aamax) aamax = abs(a[i, j]);
         }
         if (aamax == 0.) solver_error_msg("hd::lu_decomp(): singular matrix.");
@@ -96,22 +96,22 @@ void lu_decomp(mdspan<double, dextents<size_t, 2>> a,
     // LU decomposition
     double sum, aamax, dum;
     int imax;
-    for (int j = 0; j <= ubound; ++j) {
+    for (size_t j = 0; j <= ubound; ++j) {
         if (j > 0) {
-            for (int i = 0; i <= j - 1; ++i) {
+            for (size_t i = 0; i <= j - 1; ++i) {
                 sum = a[i, j];
                 if (i > 0) {
-                    for (int k = 0; k <= i - 1; ++k)
+                    for (size_t k = 0; k <= i - 1; ++k)
                         sum -= a[i, k] * a[k, j];
                     a[i, j] = sum;
                 }
             }
         }
         aamax = 0.;
-        for (int i = j; i <= ubound; ++i) {
+        for (size_t i = j; i <= ubound; ++i) {
             sum = a[i, j];
             if (j > 0) {
-                for (int k = 0; k <= j - 1; ++k)
+                for (size_t k = 0; k <= j - 1; ++k)
                     sum -= a[i, k] * a[k, j];
                 a[i, j] = sum;
             }
@@ -122,7 +122,7 @@ void lu_decomp(mdspan<double, dextents<size_t, 2>> a,
             }
         }
         if (j != imax) {
-            for (int k = 0; k <= ubound; ++k) {
+            for (size_t k = 0; k <= ubound; ++k) {
                 dum = a[imax, k];
                 a[imax, k] = a[j, k];
                 a[j, k] = dum;
@@ -133,7 +133,7 @@ void lu_decomp(mdspan<double, dextents<size_t, 2>> a,
         if (j != ubound) {
             if (a[j, j] == 0.) a[j, j] = TINY;
             dum = 1. / a[j, j];
-            for (int i = j + 1; i <= ubound; ++i)
+            for (size_t i = j + 1; i <= ubound; ++i)
                 a[i, j] *= dum;
         }
     }
@@ -172,22 +172,22 @@ void lu_backsubs(mdspan<double const, dextents<size_t, 2>> a,
     int ll;
 
     int ii = -1; // never occurring index as indicator for first loop
-    for (int i = 0; i <= ubound; ++i) {
+    for (size_t i = 0; i <= ubound; ++i) {
         ll = perm[i];
         sum = b[ll];
         b[ll] = b[i];
         if (ii != -1) {
-            for (int j = ii; j <= i - 1; ++j)
+            for (size_t j = ii; j <= i - 1; ++j)
                 sum -= a[i, j] * b[j];
         }
         else if (sum != 0.) ii = i;
         b[i] = sum;
     }
 
-    for (int i = ubound; i >= 0; --i) {
+    for (size_t i = ubound; i >= 0; --i) {
         sum = b[i];
         if (i < ubound) {
-            for (int j = i + 1; j <= ubound; ++j)
+            for (size_t j = i + 1; j <= ubound; ++j)
                 sum -= a[i, j] * b[j];
         }
         b[i] = sum / a[i, i];
